+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * Theme development related functions in CustomTheme class.
+* Disable unvanted Gutenberg SVG injection as theme default.
 
 ## 0.5.6 (2023-09-11)